+++ conversionElements.h
@@ -14,7 +14,6 @@
 #ifndef __CONVERSIONELEMENTS_H__
 #define __CONVERSIONELEMENTS_H__
 
-
 #include <stdlib.h>
 #include <math.h>
 #include <SDL2/SDL.h>
@@ -27,7 +26,6 @@ extern int windows_Height;
 extern int windows_Width;
 extern SDL_Renderer* renderer;
 extern int quit;
-
 extern int deplacX;
 extern int deplacY;
 extern float zoom;
@@ -90,26 +88,26 @@ float distanceY(float y1, float y2);
 float distanceX(float x1, float x2);
 
 /**
- * \fn This method allows to normalize coordinates functions of the vector size
- * @param  a      float Coordinate x1
- * @param  b      float Coordinate x2
- * @param  length float size of vector
- * @return        the size of the normalize vector
- */
+* \fn This method allows to normalize coordinates functions of the vector size
+* @param  a      float Coordinate x1
+* @param  b      float Coordinate x2
+* @param  length float size of vector
+* @return        the size of the normalize vector
+*/
 float normalize(float a, float b,float length);
 
 /**
- * This method calculates the angle between two vectors.
- * The returned value is in degrees
- *
- * @param  ax float X coordinate of point a
- * @param  ay float Y coordinate of point a
- * @param  bx float X coordinate of point b
- * @param  by float Y coordinate of point b
- * @param  cx float X coordinate of point c
- * @param  cy float Y coordinate of point c
- * @return    float : The angle between the two vectors AB and BC (degree)
- */
+* This method calculates the angle between two vectors.
+* The returned value is in degrees
+*
+* @param  ax float X coordinate of point a
+* @param  ay float Y coordinate of point a
+* @param  bx float X coordinate of point b
+* @param  by float Y coordinate of point b
+* @param  cx float X coordinate of point c
+* @param  cy float Y coordinate of point c
+* @return    float : The angle between the two vectors AB and BC (degree)
+*/
 float angle(float ax, float ay, float bx,float by,float cx,float cy);
 
 /**
@@ -135,30 +133,30 @@ Node* distanceToBounds(Bounds *b, Node* n);
 float distanceLatLon(float lat1, float lon1, float lat2, float lon2);
 
 /**
- * \fn This method calculates the points to draw the full polygon between two lines when is not extremite
- * \param float b The known point center
- * \param float e The new point
- * \param int signe : This shows whether one is added or subtracted depending on the given angle
- * \param Sint16 tab : The table points to draw the polygon
- * \return Sint16 The table points to draw the polygon
+* \fn This method calculates the points to draw the full polygon between two lines when is not extremite
+* \param float b The known point center
+* \param float e The new point
+* \param int signe : This shows whether one is added or subtracted depending on the given angle
+* \param Sint16 tab : The table points to draw the polygon
+* \return Sint16 The table points to draw the polygon
 */
 Sint16* midle(float b, float e, int signe, Sint16 tab[4]);
 
 /**
- * \fn boxes in the table 1 < -3 , 2 ​​> 4, 3 <0 , 4 < -0 . This puts 0 in box 3 and 4
- * \param Sint16 tab[4] The table change
- * \return The modified table
+* \fn boxes in the table 1 < -3 , 2 ​​> 4, 3 <0 , 4 < -0 . This puts 0 in box 3 and 4
+* \param Sint16 tab[4] The table change
+* \return The modified table
 */
 Sint16* swap(Sint16 tab[4]);
 
 /**
- * \fn This method calculates the points to draw the full polygon between two lines for the extremites
- * \param float a The known point center
- * \param float d The new point
- * \param float e The new point
- * \param Sint16 tab : The table points to draw the polygon
- * \param signe int it is the order
- * \return Sint16 The table points to draw the polygon
+* \fn This method calculates the points to draw the full polygon between two lines for the extremites
+* \param float a The known point center
+* \param float d The new point
+* \param float e The new point
+* \param Sint16 tab : The table points to draw the polygon
+* \param signe int it is the order
+* \return Sint16 The table points to draw the polygon
 */
 Sint16* extremite(float a, float d, float e, Sint16 tab[4],int signe, int extremite);
 
@@ -184,29 +182,29 @@ float miseAEchelleX(float x, float y,int size);
 float miseAEchelleY(float x, float y,int size);
 
 /**
- * This function calculates the offset of each point zonction zoom
- * @param  refX     The reference point x
- * @param  refY     The reference point y
- * @param  decalage The current decalage
- * @param  flag     Whether one or +zoom -zoom
- * @return          The new decalage
- */
-int calculateZoom(int refX, int refY,int decalage,int flag);
+* This function calculates the offset of each point zonction zoom
+* @param  refX     int The reference point x
+* @param  refY     int The reference point y
+* @param  decalage float The current decalage
+* @param  flag     int Whether one or +zoom -zoom
+* @return          float The new decalage
+*/
+float calculateZoom(int refX, int refY,float decalage,int flag);
 
 /**
- * According initialising the bottom of the window
- * @param  red      int color R
- * @param  blue     int color G
- * @param  green    int color B
- * @param  oppacity int oppacity color
- * @return          int Lets see if everything went well
- */
+* According initialising the bottom of the window
+* @param  red      int color R
+* @param  blue     int color G
+* @param  green    int color B
+* @param  oppacity int oppacity color
+* @return          int Lets see if everything went well
+*/
 int colorBackground(int red, int blue, int green,int oppacity);
 
 /**
- * According initialising the bottom of the window by default
- * @return int Lets see if everything went well
- */
+* According initialising the bottom of the window by default
+* @return int Lets see if everything went well
+*/
 int colorBackgroundDefault();
 
 #endif /* __CONVERSIONELEMENTS_H__ */
